==> work_notebooks/2.making a bow.ipynb <==
Take a random sample from the data and write it to a new file to avoid loading the complete data set each time

import pandas as pd
import numpy as np

df = pd.read_csv(r'data\processed\reviews_processed.csv')

df = df.sample(n = 100000)

df.to_csv(r'data\samples\sample.csv', index=False)

Write an algorithm to create a list of the unique words in all the reivews in the sample and the count of each word.

test = pd.read_csv(r'data\samples\sample.csv', nrows=600)

wc = pd.DataFrame(columns = ['word','count'])

test.dropna(inplace=True)

a = 0

for row, entry in test.iterrows():
    for word in entry['review'].split():
        if word in wc['word'].values:
            wc.loc[wc['word']==word,'count']+=1
        else:
            wc.loc[len(wc)] = [word,1]
print (wc)

The algorithm above, while functional, was too slow to run for big data sets. I optimized it by reducing the number of operations as shown in the code below, but it was still taking too long when scaled to tens of thousands of reviews. So, in the next parts I used Scikit Learn's optimized text vectorizers to speed up tokenization. 

This is faster but not fast enough:

test = pd.read_csv(r'data\samples\sample.csv', nrows=600)

wc = pd.DataFrame(columns = ['word','count'])

test.dropna(inplace=True)

a = 0

for row, entry in test.iterrows():
    for word in entry['review'].split():
        try:
            index = wc.loc[wc['word'] == word].index[0]
            wc['count'].iloc[index] += 1
        except:
            wc.loc[len(wc)] = [word, 1]
     
print (wc)
