==> work_notebooks/3.vectorized bow on initial sample.ipynb <==
import pandas as pd
import numpy as np


df = pd.read_csv(r'data\processed\reviews_processed.csv')

df = df.sample(n = 1000000)

df.to_csv(r'data\samples\sample.csv', index=False)
df.dropna(inplace=True)

reviews = df['review'].to_list()
sentiment = df['sentiment'].to_list()

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score


vectorizer = CountVectorizer(analyzer='word', ngram_range=(1, 2), max_features=3)

features = vectorizer.fit_transform(reviews)
features_nd = features.toarray()

X_train, X_test, Y_train, Y_test = train_test_split(
                features_nd,
                sentiment,
                train_size=0.8,
                random_state=1512)

log_model = LogisticRegression()
log_model = log_model.fit(X=X_train, y=Y_train)

y_pred = log_model.predict(X_test)


print(accuracy_score(Y_test, y_pred))

ratio = sentiment.count(False)/len(sentiment)
print(ratio)

feature_names = vectorizer.get_feature_names()
feature_names
