==> work_notebooks/4.bow on balanced sample.ipynb <==
import pandas as pd
import numpy as np


df = pd.read_csv(r'data\processed\reviews_processed.csv')

positive = df.loc[df['sentiment'] == True]
negative = df.loc[df['sentiment'] == False]



positive.to_csv(r'data\processed\positive_processed.csv', index=False)
negative.to_csv(r'data\processed\negative_processed.csv', index=False)

# Randomly sample 1mil  positive and negative reviews each
positive_sample = positive.sample(n=1000000)
negative_sample = negative.sample(n=1000000)

# Write samples to CSV for faster access later
positive_sample.to_csv(r'data\samples\positive_sample.csv', index=False)
negative_sample.to_csv(r'data\samples\negative_sample.csv', index=False)

import pandas as pd
import numpy as np

positive_sample = pd.read_csv(r'data\samples\positive_sample.csv')
negative_sample = pd.read_csv(r'data\samples\negative_sample.csv')

# Create lists of reviews and target sentiment values
pos_rev = positive_sample['review'].astype(str).to_list()
pos_trg = positive_sample['sentiment'].to_list()

neg_rev = negative_sample['review'].astype(str).to_list()
neg_trg = negative_sample['sentiment'].to_list()


# Concatenate review lists and target lists
reviews = pos_rev + neg_rev
targets = pos_trg + neg_trg

test = negative_sample[negative_sample['review'].str.contains('[a-z]', na=False)]
print(test.head(10))

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import accuracy_score

vectorizer = CountVectorizer(ngram_range=(1,2), max_features=200)
features = vectorizer.fit_transform(reviews)

features_nd = features.toarray()

X_train, X_test, y_train, y_test = train_test_split(features_nd, targets)

clf = LogisticRegression().fit(X_train, y_train)
y_pred = clf.predict(X_test)

print(accuracy_score(y_test,y_pred))


test = vectorizer.transform(['this was worth the money'])
print(clf.predict(test))
